requirements.txt
Flask==1.0.2
redis==3.2.1
jsonschema==3.0.1
boto==2.49.0
PyJWT==1.7.1
flask-cors==3.0.8
flask-sqlalchemy==2.4.0
flask-and-redis==1.0.0
pycryptodome==3.8.2
requests==2.22.0
flask-iniconfig==0.1.0
sqlalchemy-utils==0.33.9
mysql-connector-python==8.0.19
